Definging dtype in the array

+++ jsbgym/tasks.py
@@ -232,14 +232,14 @@ class FlightTask(Task, ABC):
         ...
 
     def get_state_space(self) -> gym.Space:
-        state_lows = np.array([state_var.min for state_var in self.state_variables])
-        state_highs = np.array([state_var.max for state_var in self.state_variables])
-        return gym.spaces.Box(low=state_lows, high=state_highs, dtype=np.float64)
+        state_lows = np.array([state_var.min for state_var in self.state_variables], dtype=np.float64)
+        state_highs = np.array([state_var.max for state_var in self.state_variables], dtype=np.float64)
+        return gym.spaces.Box(low=state_lows, high=state_highs)
 
     def get_action_space(self) -> gym.Space:
-        action_lows = np.array([act_var.min for act_var in self.action_variables])
-        action_highs = np.array([act_var.max for act_var in self.action_variables])
-        return gym.spaces.Box(low=action_lows, high=action_highs, dtype=np.float64)
+        action_lows = np.array([act_var.min for act_var in self.action_variables], dtype=np.float64)
+        action_highs = np.array([act_var.max for act_var in self.action_variables], dtype=np.float64)
+        return gym.spaces.Box(low=action_lows, high=action_highs)
 
 
 class Shaping(enum.Enum):
